+++ src/index.js
@@ -17,6 +17,8 @@ export { default as ListItem } from '../src/components/ListItem/ListItem';
 export { default as VirtualList } from '../src/components/VirtualList/VirtualList';
 export { default as InfiniteScroll } from '../src/components/InfiniteScroll/InfiniteScroll';
 export { default as UploadImage } from '../src/components/UploadImage/UploadImage';
+export { default as TextArea } from '../src/components/TextArea/TextArea';
+export { default as ColorPicker } from '../src/components/ColorPicker/ColorPicker';
 
 export { default as ThemeProvider } from '../src/utils/theme';